Add ApiUserService and TokenService
