+++ bank-of-m-and-d-fe/src/ApiService/ApiServiceAccounts.js
@@ -1,29 +1,28 @@
-export async function GetAllAccounts() {
-    const tokenFromStorage = localStorage.getItem('bearerToken');
-    const token = `Bearer ${tokenFromStorage}`;
+import { GetToken } from "../TokenService/TokenService";
 
-    return await fetch('https://localhost:55741/api/Account/all', {
-      headers: {
-          'Authorization': token,
-      }
-    });
+export async function GetAllAccounts() {
+  const token = GetToken();
+
+  return await fetch('https://localhost:55741/api/Account/all', {
+    headers: {
+        'Authorization': token,
+    }
+  });
 }
 
 export async function GetAccountById(acId) {
-  const tokenFromStorage = localStorage.getItem('bearerToken');
-      const token = `Bearer ${tokenFromStorage}`;
+  const token = GetToken();
 
-      const accountUrl = `https://localhost:55741/api/Account/${acId.toString()}`;
-      return await fetch(accountUrl, {
-        headers: {
-            'Authorization': token,
-        }
-      });
+  const accountUrl = `https://localhost:55741/api/Account/${acId.toString()}`;
+  return await fetch(accountUrl, {
+    headers: {
+        'Authorization': token,
+    }
+  });
 }
 
 export async function PostNewAccount(data) {
-  const tokenFromStorage = localStorage.getItem('bearerToken');
-  const token = `Bearer ${tokenFromStorage}`;
+  const token = GetToken();
 
   return await fetch('https://localhost:55741/api/Account', {
       method: 'POST',
@@ -36,15 +35,14 @@ export async function PostNewAccount(data) {
 }
 
 export async function DeleteAccount(data) {
-  const tokenFromStorage = localStorage.getItem('bearerToken');
-      const token = `Bearer ${tokenFromStorage}`;
+  const token = GetToken();
 
-      return await fetch('https://localhost:55741/api/Account', {
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-          'Authorization': token,
-        },
-        body: JSON.stringify(data),
-      });
+  return await fetch('https://localhost:55741/api/Account', {
+    method: 'DELETE',
+    headers: {
+      'Content-Type': 'application/json',
+      'Authorization': token,
+    },
+    body: JSON.stringify(data),
+  });
 }
